refactor(trusty-mpm): split inproject.rs to satisfy 500-SLOC production cap
Fixes a live 500-SLOC production cap violation on `main`:
`crates/trusty-mpm/src/daemon/managed_routes/inproject.rs` was 510 SLOC
and NOT allowlisted, so `scripts/check_line_cap.sh` (CI-gated) failed on
current `main`. The file grew past the cap after PR #1896 (closes
#1805/#1807) landed the old-layout migration logic plus its unit tests.

What moved where (pure refactor, zero behavior change):
- The inline `#[cfg(test)] mod tests { … }` block (9 tests, ~260 SLOC)
  moved verbatim into a sibling `inproject/tests.rs`, declared from
  inproject.rs as `#[cfg(test)] mod tests;`. This mirrors the existing
  `front_gate.rs` + `front_gate/tests.rs` pattern already used in this
  same directory.
- No production logic changed: all public API (`REPOS_ROOT_ENV`,
  `repos_root`, `repos_root_from`, `base_clone_path`, `ensure_base_clone`,
  `create_session_worktree`, `ensure_worktrees_gitignored`,
  `get_origin_url`, `try_inproject_spawn`) and private items
  (`migrate_old_layout_aside`, `OLD_LAYOUT_BACKUP_SUFFIX`) are unchanged
  and reachable identically. Test bodies moved byte-for-byte (no dedent)
  to preserve `\`-continued assertion-message string literals exactly.

Result:
- inproject.rs:       252 SLOC (production, under 500 cap)
- inproject/tests.rs: 257 SLOC (test file, under 1500 cap)
- `bash scripts/check_line_cap.sh` → 0 violations
- `cargo test -p trusty-mpm` → all pass (same 9 inproject tests, same
  module path `inproject::tests::*`)
- `cargo clippy -p trusty-mpm --all-targets -- -D warnings` → clean
- `cargo fmt --check` → clean
